+++ graylog2-web-interface/src/views/components/searchbar/SearchBarForm.test.tsx
@@ -21,10 +21,10 @@ import { AbsoluteTimeRange } from 'views/logic/queries/Query';
 import type { FormikValues } from 'views/Constants';
 
 import SearchBarForm from './SearchBarForm';
-import AbsoluteTimeRangeSelector from './date-time-picker/AbsoluteTimeRangeSelector';
+import TabAbsoluteTimeRange from './date-time-picker/TabAbsoluteTimeRange';
 
 describe('SearchBarForm', () => {
-  describe('with AbsoluteTimeRangeSelector', () => {
+  describe('with TabAbsoluteTimeRange', () => {
     it('renders', () => {
       const initialValues: FormikValues & { timerange: AbsoluteTimeRange } = {
         limitDuration: 0,
@@ -37,7 +37,7 @@ describe('SearchBarForm', () => {
       const { asFragment } = render(
         <SearchBarForm onSubmit={() => {}}
                        initialValues={initialValues}>
-          <AbsoluteTimeRangeSelector originalTimeRange={originalTimeRange} currentTimeRange={initialValues.timerange} />
+          <TabAbsoluteTimeRange originalTimeRange={originalTimeRange} currentTimeRange={initialValues.timerange} />
         </SearchBarForm>,
       );
