+++ graylog2-web-interface/src/views/components/searchbar/date-time-picker/TabKeywordTimeRange.test.tsx
@@ -22,23 +22,23 @@ import asMock from 'helpers/mocking/AsMock';
 
 import ToolsStore from 'stores/tools/ToolsStore';
 
-import OriginalKeywordTimeRangeSelector from './KeywordTimeRangeSelector';
+import OriginalTabKeywordTimeRange from './TabKeywordTimeRange';
 
 jest.mock('stores/tools/ToolsStore', () => ({}));
 
-const KeywordTimeRangeSelector = ({ defaultValue, ...props }: { defaultValue: string } & React.ComponentProps<typeof KeywordTimeRangeSelector>) => (
+const TabKeywordTimeRange = ({ defaultValue, ...props }: { defaultValue: string } & React.ComponentProps<typeof TabKeywordTimeRange>) => (
   <Formik initialValues={{ nextTimeRange: { type: 'keyword', keyword: defaultValue } }}
           onSubmit={() => {}}
           validateOnMount>
     <Form>
-      <OriginalKeywordTimeRangeSelector {...props as React.ComponentProps<typeof KeywordTimeRangeSelector>} />
+      <OriginalTabKeywordTimeRange {...props as React.ComponentProps<typeof TabKeywordTimeRange>} />
     </Form>
   </Formik>
 );
 
 jest.mock('logic/datetimes/DateTime', () => ({ fromUTCDateTime: (date) => date }));
 
-describe('KeywordTimeRangeSelector', () => {
+describe('TabKeywordTimeRange', () => {
   beforeEach(() => {
     ToolsStore.testNaturalDate = jest.fn(() => Promise.resolve({
       from: '2018-11-14 13:52:38',
@@ -73,13 +73,13 @@ describe('KeywordTimeRangeSelector', () => {
   };
 
   it('renders value passed to it', async () => {
-    const { getByDisplayValue } = await asyncRender(<KeywordTimeRangeSelector defaultValue="Last hour" />);
+    const { getByDisplayValue } = await asyncRender(<TabKeywordTimeRange defaultValue="Last hour" />);
 
     expect(getByDisplayValue('Last hour')).not.toBeNull();
   });
 
   it('calls onChange if value changes', async () => {
-    const { getByDisplayValue } = await asyncRender(<KeywordTimeRangeSelector defaultValue="Last hour" />);
+    const { getByDisplayValue } = await asyncRender(<TabKeywordTimeRange defaultValue="Last hour" />);
     const input = getByDisplayValue('Last hour');
 
     await changeInput(input, 'last year');
@@ -88,13 +88,13 @@ describe('KeywordTimeRangeSelector', () => {
   it('calls testNaturalDate', async () => {
     expect(ToolsStore.testNaturalDate).not.toHaveBeenCalled();
 
-    await asyncRender(<KeywordTimeRangeSelector defaultValue="Last hour" />);
+    await asyncRender(<TabKeywordTimeRange defaultValue="Last hour" />);
 
     expect(ToolsStore.testNaturalDate).toHaveBeenCalledWith('Last hour');
   });
 
   it('sets validation state to error if initial value is empty', async () => {
-    const { container } = render(<KeywordTimeRangeSelector defaultValue=" " />);
+    const { container } = render(<TabKeywordTimeRange defaultValue=" " />);
 
     await waitFor(() => expect(findValidationState(container)).toEqual('error'));
   });
@@ -102,7 +102,7 @@ describe('KeywordTimeRangeSelector', () => {
   it('sets validation state to error if parsing fails initially', async () => {
     ToolsStore.testNaturalDate = () => Promise.reject();
 
-    const { container } = render(<KeywordTimeRangeSelector defaultValue="invalid" />);
+    const { container } = render(<TabKeywordTimeRange defaultValue="invalid" />);
 
     await waitFor(() => expect(findValidationState(container)).toEqual('error'));
   });
@@ -110,7 +110,7 @@ describe('KeywordTimeRangeSelector', () => {
   it('sets validation state to error if parsing fails after changing input', async () => {
     ToolsStore.testNaturalDate = () => Promise.reject();
 
-    const { container, getByDisplayValue } = render(<KeywordTimeRangeSelector defaultValue="last week" />);
+    const { container, getByDisplayValue } = render(<TabKeywordTimeRange defaultValue="last week" />);
     const input = getByDisplayValue('last week');
 
     await changeInput(input, 'invalid');
@@ -119,7 +119,7 @@ describe('KeywordTimeRangeSelector', () => {
   });
 
   it('resets validation state if parsing succeeds after changing input', async () => {
-    const { container, getByDisplayValue } = render(<KeywordTimeRangeSelector defaultValue="last week" />);
+    const { container, getByDisplayValue } = render(<TabKeywordTimeRange defaultValue="last week" />);
     const input = getByDisplayValue('last week');
 
     await changeInput(input, 'last hour');
@@ -129,13 +129,13 @@ describe('KeywordTimeRangeSelector', () => {
 
   it('does not show keyword preview if parsing fails', async () => {
     ToolsStore.testNaturalDate = () => Promise.reject();
-    const { queryByText } = await asyncRender(<KeywordTimeRangeSelector defaultValue="invalid" />);
+    const { queryByText } = await asyncRender(<TabKeywordTimeRange defaultValue="invalid" />);
 
     expect(queryByText('Preview:')).toBeNull();
   });
 
   it('does not show keyword preview if parsing fails after changing input', async () => {
-    const { getByDisplayValue, queryByText } = await asyncRender(<KeywordTimeRangeSelector defaultValue="last week" />);
+    const { getByDisplayValue, queryByText } = await asyncRender(<TabKeywordTimeRange defaultValue="last week" />);
 
     asMock(ToolsStore.testNaturalDate).mockImplementation(() => Promise.reject());
     const input = getByDisplayValue('last week');
@@ -146,7 +146,7 @@ describe('KeywordTimeRangeSelector', () => {
   });
 
   it('shows error message if parsing fails after changing input', async () => {
-    const { getByDisplayValue, queryByText } = await asyncRender(<KeywordTimeRangeSelector defaultValue="last week" />);
+    const { getByDisplayValue, queryByText } = await asyncRender(<TabKeywordTimeRange defaultValue="last week" />);
 
     asMock(ToolsStore.testNaturalDate).mockImplementation(() => Promise.reject());
     const input = getByDisplayValue('last week');
